move terratest files
