+++ test/examples_complete_test.go
@@ -9,7 +9,7 @@ import (
 
 func TestExamplesComplete(t *testing.T) {
 	t.Parallel()
-	tempFolder := teststructure.CopyTerraformFolderToTemp(t, "../..", "examples/complete")
+	tempFolder := teststructure.CopyTerraformFolderToTemp(t, "..", "examples/complete")
 	terraformOptions := &terraform.Options{
 		TerraformDir: tempFolder,
 		Upgrade:      false,
